Upgrade to Gradle 7.0 in the hopes that it will support OpenJDK 16.
I pinned said hopes on this StackOverflow post:
https://stackoverflow.com/a/67955806

The upgrade had unexpected benefits.  The new version of Gradle clearly
indicated that the compileJava task was failing due to not finding the source,
so I moved app/src/ to the top-level, and lo!  java-pong-1.2.1.jar now has
Hello.class in it!

./gradlew run fails for me, but perhaps it will succeed for others.